Restructuring of the fix code, moved to the context
